Reorder scripts
Run hostname and timezone first - they are fast, don't depend on network
and we want to change the hostname as soon as possible. After that, set
up network, which may take a few seconds, but let's go online early.
Then create the user and set up SSH keys. Growing the FS may take a
longer time, so run it the last.